+++ src/login/mod.rs
@@ -2,3 +2,4 @@ pub mod github;
 pub mod gitlab;
 pub mod tools;
 pub mod google;
+pub mod provider;
